feat(manager-react-components): migrate breadcrumb component
ref: #MANAGER-18707
