+++ packages/manager-react-components/package.json
@@ -109,7 +109,7 @@
     "@ovh-ux/manager-react-shell-client": "^0.8.7",
     "@ovhcloud/ods-components": "^18.6.2",
     "@ovhcloud/ods-react": "^19.0.1",
-    "@ovhcloud/ods-themes": "^18.0.1",
+    "@ovhcloud/ods-themes": "^19.0.1",
     "date-fns": "~4.1.0",
     "i18next": "^23.8.2",
     "react": "^18.2.0",
