+++ packages/manager-react-components/src/components/breadcrumb/Breadcrumb.props.ts
@@ -1,0 +1,8 @@
+export interface BreadcrumbProps {
+  /** application name define in the shell */
+  appName: string;
+  /** root label step */
+  rootLabel: string;
+  /** hides app name from breadcrumb */
+  hideRootLabel?: boolean;
+}
