+++ packages/manager-react-components/src/components/index.ts
@@ -1,6 +1,6 @@
 export * from './action-banner';
 export * from './redirection-guard';
-export * from './breadcrumb/breadcrumb.component';
+export * from './breadcrumb/Breadcrumb.component';
 
 export * from './clipboard';
 export * from './container';
